move tests to userauth
